db: remove EXECUTE usage validations
This patch removes redundant validations when using the dynamic sql
EXECUTE keyword.
Those validations caused some functions like the
fn_db_change_column_type to reject valid type as DECIMAL and swallow
the error resulting in commands that were not executed on the database.

The EXECUTE command now throws detailed errors if the syntax given to
the EXECUTE command is not valid.

This patch also adds a fn_db_change_column_null function to switch a
column to allow/disallow NULL values instead of the wrong usage of
fn_db_change_column_type function for this purpose that never worked but
swallowed the error.

Related scripts were fixed as well one of them was in the pre-upgrade
script that always runs, the others were added as a new upgrade scripts.

Change-Id: Ia96ddef7f7450fa180a69f543cbc484d6adf0ee4
Allow-db-upgrade-script-changes: Yes
Signed-off-by: emesika <emesika@redhat.com>
Signed-off-by: Martin Perina <mperina@redhat.com>