engine: Move CPU pinning policy field
This patch moves the CPU Pinning Policy field
to the Resource Allocation tab of the VM
popup dialog. This means the field is not
related only to NUMA anymore, but it is a single place
for all CPU Pinning Policies. Therefore a new policy
called 'Manual' was added which allows to
set the manual pinning string.

Adding the MANUAL policy to cpu policies also
requires a new constraint on the vm_static table
to ensure that the cpu_pinning is set only for
manual cpu policies.

This patch also updates the REST so that
it is still possible to set/remove the cpu pinning
without defining the manual/none cpu pinning policy.

Change-Id: I7e2b76b718c23a9043ec61b387fa00ed6488c6f7
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>