core: Fix casting to Manage Block Disk on copy template disk
When trying to copy a template disk from/to a Managed Block Storage
domain, the current code treats the disk object as a DiskImage,
although its type is ManagedBlockStorageDisk.

In case of a template disk, unlike a regular disk, the copy operation
doesn't create a new disk but maps the new domain to the existing
image on the image_storage_domain_map in the DB.
Therefore, in case of copying a template disk to/from a managed block
domain, this casting fails: (ManagedBlockStorageDisk) diskDao.get().

This patch fixes this issue by creating a managed block disk based
on the disk image parameters, and add the missing values needed by the
ManagedBlockStorageDisk disk type.

Change-Id: I7a29f072a99cb165b18352b384a4c8332f15a0c4
Bug-Url: https://bugzilla.redhat.com/2006745
Signed-off-by: Shani Leviim <sleviim@redhat.com>