core: Fix duplicate DB save when copying an MBS template disk
When trying to copy a template disk from/to Managed Block Storage
domain, currently the operation fails, due to trying to save the same
image again.
Since the template disk already exists in the 'images'
(and 'base_disks') DB tables, trying to save again the disk will fail
with 'duplicate key' validation.
In case of template disks, updating the DB should be performed by
adding the additional storage domain mapping in the
'image_storage_domain_map' DB table.
Since this part is being handled in CopyImageGroupCommand, we can skip
the 'saveDisk' in AddManagedBlockStorageDiskCommand.

Change-Id: I63d7276ce7f5fc0baa66cb7fd6f273f39ad14a80
Bug-Url: https://bugzilla.redhat.com/2006745
Signed-off-by: Shani Leviim <sleviim@redhat.com>