core: wait for bitmaps to be added
Currently, the backup phase is changed to READY right after starting to
add bitmaps. However, adding bitmaps may have not finished at this point
yet, thus the backup is not actually ready.

When adding bitmaps we need to wait until all commands are complete.

This patch introduces two new phases to backups, ADDING_BITMAPS and
WAITING_FOR_BITMAPS.
These are meant to be internal phases that will be used only in cold backup
to wait for the commands adding volume bitmaps.
It also renames the runColdVmBackup to startVolumeBitmaps to match the
phase's name.

The new phases, will be reported as INITIALIZING to
the API user.

Bug-Url: https://bugzilla.redhat.com/2027260
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>
Change-Id: I529ef9f63571281513e31a65e7c4b48dda7a3ba8