core: change backup phase after starting scratch disk removal
Backup phase may be set to READY before RemoveScratchDisksCommand has
finished its execution phase, leaving a small window where a shared lock
on the VM is still held, yet the backup appears complete.
While RemoveScratchDisksCommand is not an operation we need to wait for
to complete the backup, we can have the lock it uses release before
changing the phase.

This patch sets the backup phase after initiating scratch disk removal
(the operation may still be running, but the lock should be released).

Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>
Change-Id: Iae62ec0c286d2cf9f24a29bcfb2c7f055eac2561