core: define default display
When importing a VM, the DefaultDisplayType element. Or, for a template
the default_display_type element might be missing. In such case the
entity will set into the default display type - QXL. When this is not
the real video device, the VM will be unable to run, having wrong specs
on the device.

Now, when importing a VM from non-ovirt origin, we update the static
data based on the virt-v2v and OVF reader after the convertion. In the
UI we will indicate the video type as "Not available", as by default it
would be QXL but may change later on with virt-v2v output.

Change-Id: I162418a05e30c2485d885be43dfd9c97eadc902c
Bug-Url: https://bugzilla.redhat.com/2027424
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>