webadmin: Add Storage Domain selection change handling
To fully support plugin api buttons added to the Storage Domain tab,
the `fireModelChangeRelevantForActionsEvent()` event needs to be
fired when selection changes.  This enables plugin buttons to enable
and disable based on the selected storage domains.

Updates to action buttons, both webadmin and plugin defined, now route
through the "ModelChangeRelevantForActionsEvent".  This is the same
way the ClusterListModel handels selection change.

Change-Id: I3a1bc81bebd327d5ac106b8f39c8f8f3b92f0971
Signed-off-by: Scott J Dickerson <sdickers@redhat.com>