core: default VGA video for new VMs
Until now the default video device was QXL, and the graphics were set to
SPICE+VNC. But, in the future QXL won't be supported - both on EL9 and
if one wishes to migrate to Openshift Virtualization.

In the API, QXL is left for now. As we have only one video device and
QXL supports both VNC and SPICE inputs. On a following patch, the
default display will change in API based on whether it supported.

Change-Id: Id96b469144a87454b882744bc798aa83aa9ae715
Bug-Url: https://bugzilla.redhat.com/1976605
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>