engine: Propagate UEFI for new VMs
As UEFI is becoming a preferred way of booting
operating systems for Fedora (and RHEL 9), we
should default new clusters to UEFI instead of
BIOS.

This patch changes the backend logic of the Chipset/Firmware
auto detection so that for CL levels > 4.6, the default
will be UEFI.

It also pre-selects 'Q35 Chipset with UEFI' when editing
cluster if the version is changed to CL level > 4.6. The
user may still manually set the value to Bios.

Both changes affect only creation of new VMs from the Blank template
- existing templates and VMs are not changed.

Change-Id: Ifc6e0937ca2de65867827a5dd14aed45548fe1ab
Bug-Url: https://bugzilla.redhat.com/1997893
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>