core: Add Windows 11 support
This patch adds support of Windows 11 as guest OS.

Osinfo database was extended to accommodate requirements: number of
CPUs, minimal RAM and a mandatory TPM device. These requirements are
checked when adding/updating a VM or a template. When Windows 11 is
selected in the VM editing dialog, RAM, CPUs, BIOS type and TPM device
fields are updated automatically, if needed.

Change-Id: I7296007ed994ebe03131462bc735df77a2e1036b
Bug-Url: https://bugzilla.redhat.com/1998866
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>