restapi: Base default template TPM flag on TPM device presence
When a template is created using the REST API from a VM having TPM
enabled, the template is missing TPM.  The problem is that we always
set TPM to disabled in AddVmTemplateCommand if the given parameter is
unspecified.  This is wrong, we should base the actions on the
presence of a TPM device in the original VM.

Let's not set TPM to disabled in AddVmTemplate command and handle null
isTpmEnabled value better in VmDeviceUtils::copyVmDevices.  Instead of
having two places to add the device, one in the device presence
checking loop and another one after the loop, let's add it only after
the loop, based on the parameter value and the device presence in the
original VM.

Change-Id: Idb269679bb7825d94e079654f4cc03b60a4747bc
Bug-Url: https://bugzilla.redhat.com/2024529
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>