webadmin: fix display type for new templates
Previously, when creating a template from a vm,
the default display type was not copied correctly,
although the devices were. It resulted in the default
display type being different than the video type in
the vm device table.

The cause for this was the use of the method
ListModel:setItems(Collection<T> value, T selectedItem).
The method uses OvirtSelectionModel, but is not synchronized
with method setSelectedItem(). This results in a scenario,
where value selected via setSelectedItem() is overridden by
the value previously selected by setItems().

This patch changes the way the items in the display list are
set. Instead of setItems(items, value), we are now using
two separate methods setItems(items) and setSelectedItem(item).

Change-Id: Ideaba5f98f43957a85769a8224b9734ea7079dda
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>