core: fix fraction numa pinning
When we have strange topology that leaves us with fraction when trying
to assign the CPUs in the NUMA pinning, we may end up with split cores.
If that happens, the topology set from the engine won't fit what the
guest actually has. In addition, the memory will try to align with the
CPU amount per NUMA node.

An example is a host having 1 socket, 24 cores and 2 threads.
A VM with the topology of 1 socket, 23 cores and 2 threads trying to run
and having 4 NUMA nodes pinned to the physical NUMA nodes on that host.
Using the current algorithm CPU 33 will be on NUMA node 3 while CPU 34
will be on NUMA node 4. But, they are supposed to be on the same core.
The result will be wrong topology within the guest, as `lscpu` is
confused from this situation, leading to mismatching topology inside the
guest.

This patch will correct it, by checking the threads and assigning the
whole cores.

Change-Id: I84b17d46c51af2d86c0da164c776ecbd298fbb80
Bug-Url: https://bugzilla.redhat.com/1912967
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>