core: validate amount of sockets equal to that of numa nodes
When running a VM with 'Resize and Pin' CPU pinning policy, the algorithm
expects the host to have the same number of sockets and NUMA nodes. That
way the algorithm fits the topology and leaves the first core on each
socket free to the host.

This patch filters out hosts that don't have an equal number of sockets
and NUMA nodes when trying to run a VM with CPU pinning policy
`Resize and Pin`.

Change-Id: Icdc1ac910eb733f6edd861d0505fb90a581224b5
Bug-Url: https://bugzilla.redhat.com/1955388
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>