core: fix template measurement
If CopyImageGroupWithDataCommand is invoked by AddVmFromTemplate we may
get unwanted results, as templates can be located on multiple storage
domains, which may lead to a different size calculation than the ones
desired as we may use a copy of the template that is present on a
Storage Domain different than the one selected.

While AddVmFromTemplate selects the source Storage Domain based on the
target, CopyImageGrupWithDataCommand currently ignores it and uses
sourceImage.getStorageTypes().get(0) which is wrong, as template images
may have multiple elements in the list.

Instead, the code should determine the measurement method using the
provided image and Storage Domain.

Change-Id: I0dced08f8acc413e29a4d42aaac294e979f49850
Bug-Url: https://bugzilla.redhat.com/2034531
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>