packaging: pki: Do not create lock file before use
Revert to using ca.pem itself as the lock file, essentially reverting
[1] and [2]. [1] was done, so it seems, for a case where /etc is on nfs
without functional locking. It moved the lock file from being "static"
(created by engine-setup, once per an environment) to being "dynamic" -
located on a temporary directory which is writable by anyone and might
be cleaned. This wasn't perfect, and later improved by [2], which wasn't
perfect either, even if good enough for a long time now.

Revert all of this mess and go back to using ca.pem as the lock file.
To prevent a theoretical regression fixed by [1], simply refuse to run
if the lock file is not on a local filesystem. IMO that's a resonable
compromise. [1] does not provide details for the exact bug it fixed, and
also does not handle the case where /tmp is on nfs.

If a future change for this code tries to support /etc on nfs, it might
want to find some other way to find a "lock-safe" path. It can even have
this configurable by the user somehow. IMO that's not needed right now.

[1] Change id I1d82032def commit hash 231b5ab9
[2] Change id I89d1bee3c7 commit hash aa470f72

Change-Id: I0258dc21940a3f7e68a72f48ad35ce1367e05cb9
Bug-Url: https://bugzilla.redhat.com/1990446
Signed-off-by: Yedidyah Bar David <didi@redhat.com>