webadmin: Allow to filter inherited permissions
This patch allows to filter permissions such that only direct permissions
are listed.

The patch adds a "All" and "Direct" new buttons that controls the
permission list ("All" is the default)

The permissions list is used in all major entities as DCs, Clusters,
Hosts, VMs etc.

Pushing each of the buttons triggers the search with true/false value
for the directOnly flag passed to the query

This patch includes also refactoring of SubTabQuotaPermissionView that
was not extend AbstractSubTabPermissionsView and therefor missed this
new functionality.

The permission table column title was changed from "Inherited
Permission" to "Inherited From" since this column actually shows from
which object we got the permission if the permission is not assigned
directly on a specific object.

Change-Id: Ie2ccfe14b4aa3984a7f7b6a31f1a27d2c04b753a
Bug-Url: https://bugzilla.redhat.com/2011768
Signed-off-by: emesika <emesika@redhat.com>