webadmin: Remove listener from InstanceTypeManager
When a template is changed in UnitVmModel, two listeners
are invoked. One directly in UnitVmModel, the other in
InstanceTypeManager. Both touch graphic devices and as
a result, it might fail to load the right value of graphics
protocol (the list is not yet loaded properly by one listener
at the time the other listener is setting the selected value).

This patch removes the listener from InstanceTypeManager and
calls the InstanceTypeManager's methods directly after the
UnitVmModel's listener finishes. This is effectively done
at the end of model behaviour's build method.

Change-Id: I96e6ca631fa9c0ad2e2b6c10efb75b84ad0fe83f
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>