core: avoid creating vm-manager when vm is not in db
There is an assumption that when VmManager is instantiating, the VM exists in
the database and is therefore retrieved in the post-constrcut phase (besides
unmanaged VMs). If we instantiate VmManager when the VM doesn't exist, we get
an instance that lacks the static fields of the VM (until the VM is updated
through VmManager) and this will lead to NPE when analyzing the VM via VMs
monitoring.

This patch makes sure that after the VM goes down, VmManager is not
instantiated when the VM is missing as the VM might be removed in this phase.
Instead, the existing VmManager is retrieved with a proper null-check, in
case no VmManager instance exists for the VM at that point.

Change-Id: I0ed1db7f26102ca1ab989402c5b7aade7df69cfa
Bug-Url: https://bugzilla.redhat.com/2037216
Signed-off-by: Arik Hadas <ahadas@redhat.com>