core: A VM may be deleted when in AutoStartVmsRunner list
If a VM is deleted while being restarted, AutoStartVmsRunner does not
detect this and keeps the corresponding AutoStartVmToRestart instance in
the list. The vm field in that instance is null, causing NPE and
disrupting the auto restart process for other VMs.

This patch adds the check and removal of non-existing VMs to
AutoStartVmsRunner.

Change-Id: I0d00f40d124d1e79a4bf927f7c5c7dfb5fb0e7d2
Bug-Url: https://bugzilla.redhat.com/1944637
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>