engine: add dedicated cpu filter and maintain
This patch will hold the CpuTopology in VdsManager. Each VdsCpuUnit can
be pinned. Automatically, we pin the CPU that VDSM runs on.
The dedicated CPUs will be represented in a new PendingResource.

Once the host is initialized, the engine will look on the current
running VMs to set the CpuTopology pinned CPUs correctly. Upon shutdown,
we will clear the pinned CPUs.

A logic was added to the CpuPolicyUnit filtering, we will drop the
current pinned CPUs in order to let new VMs to run on that host, having
the right number of CPUs they can use. Also, when scheduling a dedicated
CPU VM, we check if the existing VMs would be able to schedule on that
host. If one of the condition is false, we will filter that host.

As for the CpuPinningPolicyUnit, we will check if we have the topology
data and if so, if we actually can pin the VM into the host. We wish to
have it with usage within the same socket (when the number of CPUs can
be within one socket), and the same for cores.

Change-Id: Ie258d30ce0696c79ad5432c270f4d94e6ef834ba
Bug-Url: https://bugzilla.redhat.com/1782077
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>
Signed-off-by: Saif Abu Saleh <sabusale@redhat.com>