core: fix address assignment during hot-plug of virtio-scsi disks
Disks with the VirtIO SCSI interface are assigned to multiple
VirtIO-SCSI controllers when IO threads > 1. However,
previously they were all assigned to the first VirtIO SCSI
contoller when being hot-plugged, which in some scenario led
to assigning a disk with an address that is already assigned.

This patch fixes that and also improves the logic by which we
assign the disks to controllers - we now check the number of
plugged disks rather than the overall number of disks that
are attached to the VM (including the one to be plugged).

Change-Id: Ib27bfb7690a843c7af44e5342b2fbf3e302e6c81
Bug-Url: https://bugzilla.redhat.com/2040361
Signed-off-by: Arik Hadas <ahadas@redhat.com>