webadmin: fix loading of os during import
Loading of the available operating systems now
requires to specify the cluster level version. However,
the import model did not know the current cluster's version,
so if the VM did not have custom compatibility version
specified, it failed.

This patch adds a listener so that when the cluster
changes, the underlying model is notified and the
OS list is loaded.

Change-Id: I6c87d32343616077f6fffdeafc045c37016a5b48
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>