core: Exclude pooled VMs with active console from allocation
Currently a VM is assigned and taken from the pool only by users with
UserRole permission.  A user with SuperUser role has permissions to
view/open console to a VM, but he is not assigned to it. The VM is still
marked as non-attached and that's why any user with UserRole permission
can take this VM.

For solving this, this patch makes VMs that there's a console connected
to them ineligible for assignment. This is implemented by checking
vm_dynamic.client_ip field. However, if VDSM restarts while a client is
connected, client_ip gets cleared despite the console is still open. But
this flow is rare and may be resolved later.

Change-Id: I639a72d2bfe2894517f83cb268a74dbec4f2e377
Bug-Url: https://bugzilla.redhat.com/1648985
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>