core: avoid creating vm-manager when vm is not in db part 2
The previous changes prevented instantiating VmManager while
there's no static configuration in the database in case RemoveVm
is executed right after the VM switches to Down state.

However, there might be a different and more common scenario that
is addressed by this patch:
1. We execute a polling/statistics VM monitoring cycle and get VM X
2. We power off the VM and remove it
3. The host monitoring gets into a statistics cycle and uses the
   'lastVmsList', i.e., VMs that were retireived during #1, here X
4. When we iterate the 'lastVmsList' we instantiate VmManager for
   each one and in case of X, when there's no static configuration
   in the database at this point.

This is fixed by not instantiating VmManager at #4.

Change-Id: I20e9f8528dad68a3debae8557d819525b160b008
Bug-Url: https://bugzilla.redhat.com/2037216
Signed-off-by: Arik Hadas <ahadas@redhat.com>