core: Improve image transfer phase change log
Log also the current phase, so we could detect invalid phase changes.

Don't log entity message, which may be stale since we update the message
just after logging the value.

Don't log the disk id, since it is not clear what is this id, and the
log assumes that this is an image id, while disk id should be an image
group id. Also the log tried to handle case when disk id is not
available which looks impossible.

Also quote the logged values for better readability.

Change-Id: I4d23968f1cc4764570b60283c1e408503245deb9
Signed-off-by: Nir Soffer <nsoffer@redhat.com>