core: use the available SD when importing a template
If a template existed on multiple domains, its OVF will contain
refernces to all domains, as a result image.getStorageIds()[1] will contain
all domains, and get(0) will not necessarily return the selected Storage
Domain.

This patch instead maps out the available Storage Domain for each image
and will select one of them if it is not present on the target Storage
Domain.

[1] https://github.com/oVirt/ovirt-engine/blob/9bfbf94e18620c05a2854ede689605f77d16cf3f/backend/manager/modules/bll/src/main/java/org/ovirt/engine/core/bll/exportimport/ImportVmTemplateCommand.java#L156

Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>
Bug-Url: https://bugzilla.redhat.com/2043124
Change-Id: Ib7cbe6fb28aace01d4de720fd6a210b32a6fddba