core: make 8 chars password for console
Since libvirt 8 [1], a warning appears when providing a ticket with a
password with length over than 8. Until now the extra characters were
dropped silently in QEMU. But with libvirt new handling of longer
passwords it causes the console trigger to fail. Now the password we
will provided will be in the right length.

[1] - https://github.com/libvirt/libvirt/commit/27c1d06b5bd68bdce55efff0a50a15a30cb2a96b

Change-Id: I569e7a2bbb1de6ea7fd3d84e62489acca99a8349
Bug-Url: https://bugzilla.redhat.com/2042479
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>