engine: Error message during cluster upgrade
When an update of the VM/Template had failed during the cluster
upgrade from one compatibility version to another,
the audit.log did not contain any relevant information,
in the engine.log it was hard to find because the reason
might have been logged several hundreds of lines above the
cluster upgrade failed message. The UI error dialog contained
only a generic message and no reason why the edit of a VM/Template
failed.

This patch fixes the processing of the error message returned
from the VM/Template update so that every message is processed
(so far only message matching a specific pattern) and logged
in engine.log and audit.log.

The UI error dialog now contains also the error message from
the first failed VM/Template.

Change-Id: I107067dc41c74078bb2d49c5e86c685bd42fd036
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>