core: Let's mark TPM devices as special
When TPM is enabled or disabled for next run configuration, the change
is ignored when next run configuration is applied.  When processing
the devices from the next run configuration, TPM is treated as "other"
device, VmDeviceCommonUtils#isSpecialDevice returns false for it and
then TPM is processed as an unmanaged device.  But TPM is actually a
managed device, so let's fix the problem by returning true from
VmDeviceCommonUtils#isSpecialDevice for TPM devices.

Change-Id: I99dd9ba97660d86f98d7cf7e26d4112c99ad4bcb
Bug-Url: https://bugzilla.redhat.com/1999028
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>