engine: Small Blank template fixes
The patch fixes the following issues:

a) Some fields have not been set with data from
model, although they are visible in Blank template dialog.
This was because some of the builders were deliberately
left out. Now the same builders as for the existing VM are
used.

b) If the Cirrus video type was selected, the graphics protocol
was not loaded.

c) If the Cirrus video type was selected, it was not stored in
the backend, but replaced by QXL.

d) The Bios type field has been hidden as it is not possible to
select a value for Blank template anyway. It is now consistent
with cluster field.

e) Setting fields as (non) available has been moved to
initialize() method.

f) Removed call to updateResumeBehavior() as the method requires
the cluster to be selected and if not (e.g. for Blank template),
it returns.

g) The custom compatibility version has been hidden as it makes
no sense on the Blank template

Change-Id: Ic30af467128943d825da711150397ce0306ac456
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>