core: set disk_id when downloading snapshots
Currently when downloading snapshots the disk_id would remain empty.
This may cause issues if this field is accessed, which happens when
moving a host to maintenance and will lead to NPEs.

Since snapshots in the system are always part of a disk, this patch sets
the snapshot's disk_id when transfer request is made.

Bug-Url: https://bugzilla.redhat.com/2043984
Change-Id: I28cc15d0d772d52876a35d11e7adb4660de48a90
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>