core: skip snapshots for VMs with MBS disks
When cloning a VM with an MBS disk, the method getSourceDisks() is
being used for getting and filtering the target disks for cloning.
Currently, the MBS disks are filtered by their pluggability.
During the cloning process, a VM snapshot is being created, and its
disks are marked as un-plugged and not active.
Therefore, the MBS disks are filtered out from the images' list.
When trying to remove the snapshot and retrieve the disks,
there's an NPE, and the clone operation fails.

This patch skips the snapshots' creation (and removal) in case of a
non-running VM with MBS disks.

Bug-Url: https://bugzilla.redhat.com/1982083
Signed-off-by: Shani Leviim <sleviim@redhat.com>