core: pass full scratch disk info in 'VmBackupVDSParameters'
As a preparation to the follow-up  commit, storing the whole
scratch disk object in 'VmBackupVDSParameters', so it can be
included in 'VmBackupConfigVDSCommandBase' when creating a
backup configuration for VDSM.
As part of the change, also refactored and simplified the code that
prepared the scratch disk, to have a simple & readable "ScratchDiskInfo"
class vs. using the "Pair", the latter made the code hard to follow
due to cross-class passing & usage of this Pair objects.

Signed-off-by: Pavel Bar <pbar@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1913389
Change-Id: I4841d69c72a8c82cd4b257bf7402a4574818070b