webadmin: avoid resetting host selection on upload disk
When uploading a disk from Storage > Disk > Upload, after selecting a
host different from the first one on the list, the selection immediately
resets back to the first host suggested.
By initializing the unchangeable data center and storage domain once
the model loads for the first time, we avoid the resetting.

Bug-Url: https://bugzilla.redhat.com/2041544
Signed-off-by: Shani Leviim <sleviim@redhat.com>