core: Store mediated devices as VM devices
Information about mediated devices is currently stored as a VM custom
property, in a textual form.  We need to add more mediated device
properties and this calls for using a structured form.

Let's store mediated devices as fake VM devices, one VM device for
each mdev type.  This also allows using nodisplay property per each
mdev type.

In order not to change device configuration when starting a VM, the
check for the cluster version is performed only in the place where
it's needed and is not part of the device configuration or its
retrieval itself.

This patch is a preparation patch for adding support for vGPU plugin
parameters on the host.  In a future patch, we would like to add
another device property for it.  It may become more complex, because
the plugin parameter is set per host device and we would need to
handle possibly conflicting requirements.  Nevertheless, this is a
stuff for the future patch.  Mediated device properties cleanup is
useful anyway.

Bug-Url: https://bugzilla.redhat.com/1987121