core: no gateway 0.0.0.0 should be in sync
When a user configures a 0.0.0.0 gateway, vdsm reports it back as an
empty string which engine then converts into null. Comparing null to the
configured 0.0.0.0 results in reporting the attachment as out of sync
when actually null signifies 'no gateway' as is signified by 0.0.0.0
Therefore, align the comparison predicate with this corner case.

Change-Id: I48d6039dd1044ce5997f8bb152785ab0d78b4e54
Bug-Url: https://bugzilla.redhat.com/1993016
Signed-off-by: Eitan Raviv <eraviv@redhat.com>