core: Avoid filtering MBS disks when creating a template from snapshot
Currently, when creating a template from a snapshot which has MBS disks,
the operation is being completed successfully since the MBS disks are
being omitted from the resulted template.
As that operation isn't being currently supported for MBS disks, it should be blocked.
This patch avoids filtering out the MBS disks and leads to the deserved
outcome of blocking the operation.

Bug-Url: https://bugzilla.redhat.com/2004350
Signed-off-by: Shani Leviim <sleviim@redhat.com>