core: unite 'PREPARING_SCRATCH_DISK' & 'STARTING' phases
Remove the "PREPARING_SCRATCH_DISK" phase and unite
its content with a 'STARTING' phase in order to:
1. Improve efficiency - the state is not really needed and moving
   between the phases unnecessary increases the backup overall time.
2. As a preparation to the follow-up  commit that will
   introduce the "ScratchDiskInfo" object:
  a. We don't want the new "ScratchDiskInfo" object to be
     part of the "VmBackupParameters" class (the latter is used
     to store state between backup phases).
  b. By uniting the 2 phases, we are able to create, initialize
     and use the new object at the same phase where "StartVmBackup"
     command is invoked (and "VmBackupConfigVDSCommandBase" is
     initialized during "StartVmBackupVDSCommand" initialization).

Note: the original implementation (before phase uniting) had
the new "ScratchDiskInfo" object used over 3 phases:
a. Partial initialization of the "ScratchDiskInfo" object
   in "CREATING_SCRATCH_DISKS" phase.
b. Finishing the initialization of the above object in "PREPARING_SCRATCH_DISK"
   phase (which is obviously a pretty ugly implementation).
c. Finally using the object in "STARTING" phase.

By this code refactoring to unite the 'PREPARING_SCRATCH_DISK' & 'STARTING'
phases, in addition with a small trick to not use the "ScratchDiskInfo" object
in "CREATING_SCRATCH_DISKS" phase, we will be able to avoid the need for
the object to be part of the "VmBackupParameters" class, because there will be
no need to pass this info between backup phases.

Signed-off-by: Pavel Bar <pbar@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1913389
Change-Id: I9352171b9760bbb4f3cf2827903b4c530389d725