core: refresh mdev devices when vms with run-once configuration stop
ProcessDownVm removes unmanaged devices for stateless VMs and VMs that
started in run-once mode. This created a problem later on when checking
whether the host needs to be refreshed because there we checks for the
existence of host devices that might have been removed already, like
mdev devices.

Now, we check the devices of the VMs prior to calling ProcessDownVms so
we will determine whether the host should be refreshed properly, also in
case of stateless VMs or VMs that were started in run-once mode.

Bug-Url: https://bugzilla.redhat.com/2052557
Signed-off-by: Arik Hadas <ahadas@redhat.com>