packaging: setup: editConfigContent: Support removing params
Its docstring says that params set to (python) None are removed. But
this does not work - was never used, so not noticed so far.

I want to use this in a next patch, thus current patch.

Trying to fix this using the existing code was hard, and the code was
too complex and hard to understand in my opinion, so I mostly rewrote
it. I hope the new code is easier to understand and maintain.

Change-Id: I3b30465159e0e1d894ef591100bf987e4fc1f138
Bug-Url: https://bugzilla.redhat.com/1999698
Signed-off-by: Yedidyah Bar David <didi@redhat.com>