restapi: fix vm mapper
The VM mapper was not creating DynamicCpu instance for dedicated CPUs which resulted in NPE.

This patch changes the order of filling the model - at first, if the cpu pinning policy is dynamic, the DynamicCpu is created. Then, if the dynamic pinning is numa resize and pin, the dynamic cpu topology is set.