core: fix NPE on cpuTopology
When using older hosts, the cpu_topology is not reported from the host
caps. This caused a NullPointerException when streaming over the list,
as the list is a null and not an empty list.
On the VdsCpuUnitPinningHelper, when using manual pinning and by the
CpuPolicyUnit a caution about empty CPU topology needed.

This patch will initiate the VdsManager cpuTopology, to have an empty
list instead of null. Also, when entering some flows in the helper if
the cpuTopology is empty we should return immediately.

Change-Id: I6801dbbf4c63f96061a4c16dd001fb7260c717a1
Bug-Url: https://bugzilla.redhat.com/1782077
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>