core: Add backend support for vGPU driver parameters
This patch adds support for specifying driver parameters for a vGPU
device.  It's a string to be written to the driver on the host side.
It's stored alongside with other vGPU parameters in VM metadata.

If the host doesn't support this feature, the driver parameters will
be ignored on the host.  We log a warning for cluster levels not
supporting the feature.

Driver parameters are supported only for NVidia cards.  This is not
checked anywhere, the VM will fail to start if it is attempted to set
driver parameters for other cards.

Bug-Url: https://bugzilla.redhat.com/1987121