core: Fix wrong IP comparison in NetworkConfigurator
The network configurator was comparing only string
representation of IP addresses which could lead to
trying to setup management network even when there was
already one on the host.

The problem is most obvious for the IPv6, the InetAddress
returns expanded form of address e.g. "fbc::10" becomes
"fbc:0:0:0:0:0:0:10" which compared to the original
unexpanded form fails.

To prevent this use IPAddressPredicate to compare the values,
as the predicate is capable of comparing expanded and
unexpanded form.

Signed-off-by: Ales Musil <amusil@redhat.com>
Bug-Url: https://bugzilla.redhat.com/2061904