core: move resize and pin CPUs settings
Since we introduced `dedicated` policy, we need to allocate the physical
resources even for other policies such: `manual` and `resize and pin
NUMA`. Until now the CPU topology and CPU pinning for `resize and pin
NUMA` happened after host selection by the scheduler. But in that phase
we allocate the physical resources, which result us in NPE and
potential to wrong data for the host resources.
Now, the CPU settings such as topology and CPU pinning will be generated
after host selection, when allocating the physical CPUs.

Change-Id: Iff425646825cd258a9c35c88f36cd4b66618eac5
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>