core: Support old mdev_type property in OVF reading and writing
The old mdev_type property is not supported anymore, but it may still
be present in snapshots and exported VMs.  We must be able to read it.

We must be also able to write it if we are in a cluster level of
Engines that didn't support mediated VM devices yet.  However, the VM
devices may contain more information than the mdev_type property can
represent.  We cannot do anything about it if the cluster level is
insufficient and we must perform possibly lossy conversion.  But if at
least the custom cluster level is sufficient, we don't write mdev_type
property and keep the mediated VM devices.

Bug-Url: https://bugzilla.redhat.com/1987121