engine: Warn if a passthrough device is used on the host
PCI passthrough devices should be blacklisted on the host to prevent the
host to load drivers for them. Otherwise other processes on the host may
access the devices passed to a VM guest, causing various problems.

This patch adds a list of allowed drivers for PCI passthrough to the
configuration. A warning is added to the audit log when starting a VM,
if the driver is not in the list. The list currently contains
'vfio-pci', 'pci-stub' and '' (empty string) for devices that have
blacklisted module but don't have an alternative driver assigned.

Change-Id: Ic6520ea0a35f8b7c099b47aef2de4034c96a3bc7
Bug-Url: https://bugzilla.redhat.com/1960808
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>